+++ bpf/libbpf/netfilter.bpf.c
@@ -11,7 +11,6 @@
  * netfilter framework") v6.3-rc6-1646-gfd9c663b9ad6 support.
  */
 #include "vmlinux.h"
-#include <linux/version.h>
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_endian.h>
 
@@ -20,11 +19,6 @@
 #define ETH_P_IP        0x0800
 #define ETH_P_IPV6      0x86DD
 
-#ifndef LINUX_VERSION_CODE
-#error "Not found LINUX_VERSION_CODE in linux/version.h"
-#endif
-
-#if LINUX_VERSION_CODE > KERNEL_VERSION(6, 3, 0)
 SEC("netfilter")
 int netfilter1(struct bpf_nf_ctx *ctx)
 {
@@ -38,8 +32,5 @@ int netfilter1(struct bpf_nf_ctx *ctx)
 		return NF_DROP;
 	}
 }
-#else
-#warning "only kernel >= v6.3-rc6 support BPF_PROG_TYPE_NETFILTER"
-#endif
 
 char _license[] SEC("license") = "GPL";